fix(security): domain-separate signed TRC10 metadata
- add canonical versioned and length-delimited payloads for TokenDetails and ExchangeDetails
- preserve published seven-digit legacy signatures behind strict, disjoint validation
- prevent signature replay across token and exchange metadata types
- support canonical 1–19 digit TRC10 IDs without changing protobuf wire schemas
- expand nanopb buffers for future eight-digit and longer token IDs
- update metadata signing tools and regression coverage
- document Ledger approval, re-signing, rollout, and legacy retirement TODOs

+++ src/tokens.h
@@ -20,7 +20,7 @@
 
 int verifyTokenNameID(const char *tokenId,
                       const char *tokenName,
-                      uint8_t decimals,
+                      uint32_t decimals,
                       uint8_t *signature,
                       uint8_t signatureLength);
 int verifyExchangeID(const unsigned char *exchangeValidation,
